move TxIn and TxId to core
