+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState.hs
@@ -199,7 +199,6 @@ import Cardano.Ledger.Shelley.TxBody
     Ptr (..),
     RewardAcnt (..),
     TransTxId,
-    TxIn (..),
     Wdrl (..),
     WitVKey (..),
     getRwdCred,
@@ -219,6 +218,7 @@ import Cardano.Ledger.Slot
     EpochSize (..),
     SlotNo (..),
   )
+import Cardano.Ledger.TxIn (TxIn (..))
 import Cardano.Ledger.Val ((<+>), (<->), (<×>))
 import qualified Cardano.Ledger.Val as Val
 import Cardano.Protocol.TPraos (PoolDistr (..))
