+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/EraGen.hs
@@ -59,10 +59,10 @@ import Cardano.Ledger.Shelley.Constraints (UsesPParams (..))
 import Cardano.Ledger.Shelley.LedgerState (UTxOState (..))
 import Cardano.Ledger.Shelley.PParams (ProtVer, Update)
 import Cardano.Ledger.Shelley.Rules.Utxo (UtxoEnv)
-import Cardano.Ledger.Shelley.Tx (TxId (TxId))
-import Cardano.Ledger.Shelley.TxBody (DCert, TxIn, Wdrl, WitVKey)
+import Cardano.Ledger.Shelley.TxBody (DCert, Wdrl, WitVKey)
 import Cardano.Ledger.Shelley.UTxO (UTxO)
 import Cardano.Ledger.Slot (EpochNo)
+import Cardano.Ledger.TxIn (TxId (TxId), TxIn)
 import Cardano.Protocol.TPraos.BHeader (BHeader)
 import Cardano.Slotting.Slot (SlotNo)
 import Control.State.Transition.Extended (STS (..))
