+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Shrinkers.hs
@@ -10,9 +10,10 @@ import qualified Cardano.Ledger.Core as Core
 import Cardano.Ledger.Shelley.BlockChain
 import Cardano.Ledger.Shelley.PParams
 import Cardano.Ledger.Shelley.Scripts
-import Cardano.Ledger.Shelley.Tx
-import Cardano.Ledger.Shelley.TxBody
+import Cardano.Ledger.Shelley.Tx hiding (TxIn)
+import Cardano.Ledger.Shelley.TxBody hiding (TxIn)
 import Cardano.Ledger.Slot
+import Cardano.Ledger.TxIn (TxIn)
 import Cardano.Ledger.Val ((<+>), (<->))
 import qualified Cardano.Ledger.Val as Val
 import Data.Foldable (toList)
